Rename: pr 리뷰사항(챌린지 성취 관련 dto 패키지 변경)
