+++ src/main/java/com/dnd/runus/domain/challenge/achievement/ChallengeAchievementRecord.java
@@ -1,4 +1,4 @@
-package com.dnd.runus.domain.challenge.achievement.dto;
+package com.dnd.runus.domain.challenge.achievement;
 
 public record ChallengeAchievementRecord(ChallengeAchievement challengeAchievement, PercentageValues percentageValues) {
     public ChallengeAchievementRecord(ChallengeAchievement challengeAchievement) {
