+++ src/main/java/com/dnd/runus/domain/challenge/achievement/ChallengeAchievementRepository.java
@@ -1,7 +1,5 @@
 package com.dnd.runus.domain.challenge.achievement;
 
-import com.dnd.runus.domain.challenge.achievement.dto.ChallengeAchievement;
-
 import java.util.Optional;
 
 public interface ChallengeAchievementRepository {
