+++ src/main/java/com/dnd/runus/domain/challenge/achievement/PercentageValues.java
@@ -1,3 +1,3 @@
-package com.dnd.runus.domain.challenge.achievement.dto;
+package com.dnd.runus.domain.challenge.achievement;
 
 public record PercentageValues(int achievementValue, int startValue, int endValue) {}
